+++ lib/ts/recipe/thirdpartyemailpassword/recipe.tsx
@@ -52,9 +52,9 @@ export default class ThirdPartyEmailPassword extends AuthRecipeModule<
     static instance?: ThirdPartyEmailPassword;
     static RECIPE_ID = "thirdpartyemailpassword";
 
-    emailPasswordRecipe: EmailPassword;
+    emailPasswordRecipe: EmailPassword | undefined;
 
-    thirdPartyRecipe: ThirdParty;
+    thirdPartyRecipe: ThirdParty | undefined;
 
     recipeImpl: RecipeInterface;
 
@@ -77,6 +77,8 @@ export default class ThirdPartyEmailPassword extends AuthRecipeModule<
         this.emailPasswordRecipe =
             recipes.emailPasswordInstance !== undefined
                 ? recipes.emailPasswordInstance
+                : this.config.disableEmailPassword
+                ? undefined
                 : new EmailPassword(
                       {
                           appInfo: this.config.appInfo,
@@ -102,9 +104,14 @@ export default class ThirdPartyEmailPassword extends AuthRecipeModule<
                       }
                   );
 
+        // we initialise this recipe only if the user has provided thirdparty
+        // providers.
         this.thirdPartyRecipe =
             recipes.thirdPartyInstance !== undefined
                 ? recipes.thirdPartyInstance
+                : this.config.signInAndUpFeature.providers === undefined ||
+                  this.config.signInAndUpFeature.providers.length === 0
+                ? undefined
                 : new ThirdParty(
                       {
                           appInfo: this.config.appInfo,
@@ -131,10 +138,21 @@ export default class ThirdPartyEmailPassword extends AuthRecipeModule<
     }
 
     getFeatures = (): RecipeFeatureComponentMap => {
-        const features: RecipeFeatureComponentMap = {
-            ...this.emailPasswordRecipe.getFeatures(),
-            ...this.thirdPartyRecipe.getFeatures(),
-        };
+        let features: RecipeFeatureComponentMap = {};
+
+        if (this.emailPasswordRecipe !== undefined) {
+            features = {
+                ...features,
+                ...this.emailPasswordRecipe.getFeatures(),
+            };
+        }
+
+        if (this.thirdPartyRecipe !== undefined) {
+            features = {
+                ...features,
+                ...this.thirdPartyRecipe.getFeatures(),
+            };
+        }
 
         if (this.config.signInAndUpFeature.disableDefaultImplementation !== true) {
             const normalisedFullPath = this.config.appInfo.websiteBasePath.appendPath(new NormalisedURLPath("/"));
@@ -152,6 +170,9 @@ export default class ThirdPartyEmailPassword extends AuthRecipeModule<
 
     getDefaultRedirectionURL = async (context: GetRedirectionURLContext): Promise<string> => {
         if (context.action === "RESET_PASSWORD") {
+            if (this.emailPasswordRecipe === undefined) {
+                throw new Error("Should not come here...");
+            }
             return this.emailPasswordRecipe.getDefaultRedirectionURL(context);
         } else {
             return this.getAuthRecipeModuleDefaultRedirectionURL(context);
@@ -171,6 +192,9 @@ export default class ThirdPartyEmailPassword extends AuthRecipeModule<
                 </AuthWidgetWrapper>
             );
         } else if (componentName === "resetpassword") {
+            if (this.emailPasswordRecipe === undefined) {
+                throw new Error("Should not come here...");
+            }
             return this.emailPasswordRecipe.getFeatureComponent(componentName, props);
         } else {
             return this.getAuthRecipeModuleFeatureComponent(componentName, props);